Update XenGT I/O emulation logic.
The XenGT I/O emulation code is not correctly implemented and can
race with Xen, leading to a domain_crash() of the guest. Specifically
it is always unsafe for an emulator to drive ioreq state from NONE
to anything else, and XenGT may drive it from NONE to RESP_READY
This patch is to fix this issue on XenServer, from Citrix Paul Durrant.
Corresponding bug number in bugzilla is: #837.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>
Signed-off-by: Terrence Xu <terrence.xu@intel.com>