kvmgt: rework opregion according userspace changes
Now in QEMU we have the OpRegion backed by normal memory, so
in kernel we should also initialize the region by pin user
hva and map it.

This patch also renamed the workaround indicator 'vgt_in_xen'
to 'opregion_present', since opregion is the only differece
that not handled by MPT implementation.

Signed-off-by: Xiaoguang Chen <xiaoguang.chen@intel.com>
Signed-off-by: Jike Song <jike.song@intel.com>