Fix the KVMGT reboot issue - (kvmgt part)
In KVMGT, there’s an issue that guest OS cannot reboot normally. This fix is to resolve the reboot issue in kvm.
The reason why we need these extra steps in kvm is that:
 --When rebooting a guest OS, Xen will destroy the existing qemu process and start a new one, so VGT related traps can be free correctly.
 --But in kvm, the same qemu process will be remained, so VGT traps like mmio trap and PPGTT wp pages will remain working during the reboot of the guest OS, which will cause the guest fail to reboot.

In this fix, we have two parts: kvm part and qemu part.

This patch in kvm implemented a function to unset the mmio trap area in kvm, and called to it when releasing vgt instance.

Signed-off-by: Min He <min.he@intel.com>