Fix an address calculation error for memcpy in gtt.c
The offset added to val64 should be byte-wise, but the base pointer
type is not char *. It will cause the wrong address calculation
result.

The memcpy()s are used to update guest page table. It will cause some
guest entries to be cleared to 0 unexpectedly. So far its impact is
limited. We do not have bug related to it and it was found in code
review.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>