Enable shadow command buffers
Let command scan and patch list apply to shadow command buffers, and
enable shadow command buffers.

In order to support this, the command scan basic utilities are changes.
The previous functions like cmd_val() etc. works for normal gma,
which gets gpa from guest GTT table, and then va from gpa-to-va mapping.
The access to the va address needs to call mtp API. With shadow command
buffer, the rb/bb are in reserved aperture. The access is through aperture
va directly.

The command scan for dom0 is removed. In earlier XenGT implementation
we need to noopify Dom0's display commands in command buffer when dom0
is not foreground VM. But in BDW, when execlist mode is enabled,
"mmio_flip" will be used and command buffer does not need any patching
work.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Signed-off-by: Fred Gao <fred.gao@intel.com>