vgt: remove ASSERT from cmd_parser.c
We should return an error code when some invalid command data occurs
from a VM. It can be a attack if a malicious invalid GPU command
submitted from a VM, which cause host hang.

v3. Keep ASSERT_VM at the moment. We will replace it with other
function later.

v2. Refine error message for the case 'NULL address to be patched' in
add_patch_entry

Signed-off-by: Yi Sun <yi.sun@intel.com>