remove validating empty context checking while submiting context
There is case that guest submit workloads through lite-restore, which means guest
driver kept updating tail pointer, since there is no context-done interrupt so guest
context header pointer did not get updated.
By chance the tail pointer equals to header pointer because of the not updated header pointer.
At that time the batch buffer's header and tail keep zero, and the submission was skipped.
We should delete the check for command scan can know whether there are workloads or not.

This fix bug 869.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Signed-off-by: Xiaoguang Chen <xiaoguang.chen@intel.com>