Fix kernel panic caused by element switch
We queued workload submitted from guest, and when context switch
interrupt is received, we will emulate the status change of those
submitted workloads. Normally the search of the queued workload should
be First-In-First-Out order, but we have logic to change the order in
some scenarios.

The order change is not correct. We should always follow the FIFO
order. In most cases the old code do not cause problem, but will cause
failure in below case:

Submission:

1, guest submit ctx pair A: (CTX-1, CTX-2), then is scheduled to hardware
2, guest submit ctx pair B: (CTX-2, CTX-1), then is scheduled to hardware

Interrupt:

a, CTX-1 idle-to-active
b, CTX-1 element-switch
c, CTX-2 active-to-idle

d, CTX-2 idle-to-active
e, CTX-2 element-switch
f, CTX-1 active-to-idle

VGT Emulation:

The problem is in the emulation of interrupt (b). We should find the
submission from the record "A". If we do not follow FIFO order, we
will find the record "B". Then VGT got confused that the
element_switch is reported with the second context's context_id.

V2:
 - Remove the forward_search function argument completely. (from Kevin)

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>