use page address for reserved aperture
In shadow context, previous implementation is to use the virtual
address corresponding to aperture memory, which is performance low due
to non-cachable memory access. This commit changes to use page virtual
address instead.

Before doing that, we set correct dma mask in driver so that
pci_map_page() does not allocate new pages for iova mapping.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>