Add IS_PREEMPTION check when update vring scan state.
We have Lite_Resote case, the new el_ctx ring head = last head,
and we need to scan the command buffer begin at last_scan_head.
But we have special case it have same state as Lite_Restore,
but one resubmission. For this case,
we need to use the ring head as scan head.
Bug Fix: 899

Signed-off-by: Weinan Li <weinan.z.li@intel.com>