Fix a potential QEMU crash issue
When we clear the write protection of a page, there is a small window
causing problem. Our current code is to remove the page from ioreq
rangeset, then clear the page type for trap. If there is the page
access inside the small window, it will be trapped into hypervisor,
but the page is not inside ioreq rangeset. Then the page fault will be
directed into QEMU and cause QEMU to exit.

The fix is to change the order of the two hyper-call for "clear wp".

It was found in code review, and no bugs directly related to it. But
earlier we ever met random "privcmd_fault" from QEMU. That may be
related to this problem.

v2:
 - passing constant map/unmap value in function call directly
   since we know that for sure (Kevin)

Cc: Kevin Tian <kevin.tian@intel.com>
Cc: Zhi WAng <zhi.a.wang@intel.com>

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>