vgt: do not kill when fail to resize the mempool.
Sometimes the mempool resize is not successfully when the whole system
is under memory pressure, but this failure is not fatal and will not
harm the existing mempool.
The root cause is upstream code is using kmalloc to allocate a continous
memory region for mempool element data structure. It's not easy to
acllocate a big continous memory element data structure when system under
memory pressure. But changing it to vmalloc is not realistic, as currently
some code in block subsystem will resize the poll with GFP_NOWAIT.
So we keep VM alive even the resize is fail.

The corresponding bug number in bugzilla are: #930.

Signed-off-by: Min He <min.he@intel.com>
Signed-off-by: Fred Gao <fred.gao@intel.com>
Signed-off-by: Terrence Xu <terrence.xu@intel.com>