vgt: fix bug 836 issue
V4: refined the description and added error process.

V3: refine the logic of the PTE handling.

V2: modified the description of the patch according to Kevin's advise.

It is a workaround for bug 836, and is only for bc branch. In this
patch, we added a scratch page to each domain and assign this page to
PTE when removing a PTE entry.

Signed-off-by: Xiao Zheng <xiao.zheng@intel.com>
Signed-off-by: Min He <min.he@intel.com>