remove vmfb_mapping
The gem vgtbuffer implementation hacks i915 gem object too much. A
flag "vmfb_mapping" was added into gem object, and the object binding
to gm will check the flag to call different function for vgtbuffer.

The implementation is broken in previous code rebase. It was not
investigated and is currently feature disabled. The vgtbuffer will
have a new implementation, so here we remove all the hacks in
i915_gem.c and i915_gem_gtt.c

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>

Conflicts:
	drivers/gpu/drm/i915/i915_drv.h
	drivers/gpu/drm/i915/i915_gem.c
	drivers/gpu/drm/i915/i915_gem_gtt.c