Remove the incorrect interrupt register from reset list
During the vgt initialization and reset, the cached interrupt registers
would be reset bassed on MMIO value for correspoinding imr and ier. But
for GEN8_MASTER_INTERRUPT, it doesn't have the corresponding imr and ier
registers. Current code doesn't see this and do the MMIO read operations
for GEN8_MASTER_INETERRUPT which will cause a warning print because the
accessed 'registers' are not claimed.

Signed-off-by: min <min.he@intel.com>
Signed-off-by: pei <pei.zhang@intel.com>