Adjust cmd length correctly for NOOP
It was expected that MI_NOOP command will have ID_NUMBER field all zero when
WRITE_ENABLE bit (bit 22) was not set. But we see that it was not always true.
Then GVT-g cannot use ID_NUMBER field to record the cmd buffer patch info.

Here this patch is a temporary change to avoid the miscalculation of MI_NOOP
command length. Later patch will have a more complete fix.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>