Move the extra NOOPs update before submittion to avoid dummy commands submitted
Ported from upstream i915 repo:
----
  commit 7c17d377374ddbcfb7873366559fc4ed8b296e11
  Author:Chris Wilson <chris@chris-wilson.co.uk>
  Date:  Wed, 20 Jan 2016 15:43:35 +0200
  Subject: [PATCH] drm/i915: Use ordered seqno write interrupt generation
   on gen8+ execlists
----

There are two extra NOOPs as padding to avoid lite restore of a context
with HEAD==TAIL. But there is risk submitting dummy commands to GPU,
since it write the NOOPs commands after submission, 1 or 2 unexpected
commands may be submitted to GPU ring when haven't writen in time.
Do it before submission to avoid it.

Only porting the code of writing NOOPs sequence, but not the write
sequence number command update, need to merge it during rebase.

Signed-off-by: Weinan Li <weinan.z.li@intel.com>