vgt: remove waitqueue_active() check in xen mpt module
To call waitqueue_active(), holding the queue spinlock is required,
which is missed in current code piece. The kernel thread might miss
an IO request should be processed due to the race condition between
waitqueue_active() and wake_up().

And it's also not necessary to leverage waitqueue_active() as an
optimization here. If the wait queue is really empty, wake_up()
will do nothing.

Signed-off-by: Zhi Wang <zhi.a.wang@intel.com>