SBI read cache miss issue fix
V2. change vgt_warn to vgt_dbg.

vGT will cache guest's SBI data in a local array. This cache
operation happens everytime when the SBI write occurs. So when
there is a SBI read, vGT returns the cached data to guest.
This is good if SBI read always happens after write. But it's
not always true. The guest OS may first try read, default 0
will returned, if 0 is not the value it wanted, it then write
a new value. This will cause a warning print in the first time
read.

More terrible thing is that, if the default value 0 returned in the
first time read could satisfy the guest's requirement, there won't
be a SBI write followed and vGT won't find a chance to cache the SBI
data for guest OS. For this special SBI addr, every following read
will cause a waring print. In some condition, we can see huge SBI
cache miss warning print (addr 0x60c is an example).

So I change current logic and also cache SBI data in read time.

Signed-off-by: Pei Zhang <pei.zhang@intel.com>