vgt: polish and export module parameters to debugfs/runtime
Set the permission for each module parameter, as well as
create symlinks for the module parameters which can be runtime changed
in /sys/kernel/debugfs/vgt/runtime/xxx.

Initailly, vgt use module_param_named macro to export module parameters
in /sys/module/i915/parameters/xxx, and the permissions for module parameters are coarse-grained.

In this patch, the permission for each module parameter is clearly and
carefully set (0600, 0400, etc.), and symlinks are created for some module parameters in debugfs/vgt/runtime.

Signed-off-by: Yu Xu <yu.a.xu@intel.com>
Signed-off-by: Zheng, Xiao <xiao.zheng@intel.com>