gdemon-kernel: release checkpoint
In this patch, we present gDemon, which is a concrete implementation of
Demon. gDemon reuses the guest GTT as the 1st-dimensional translation
and utilizes the system IOMMU to construct a 2nd-dimensional address
translation. gDemon dynamically switches the 2nd-dimensional page table
when the device owner switches. In order to enable guest VM to access
its own global graphics memory, such as AGP Aperture, even when the
corresponding vGPU is not current device owner, GGTT remains
partitioned, shadowed, and shifted to a dedicated address range that
PPGTTs won't map.
The evaluations show that gDemon provides up to 19.73x better
performance in the media transcoding workloads and achieves the
performance improvement of up to 17.09% and 13.73% in 2D benchmarks and
3D benchmarks, respectively, compared to gVirt. Moreover, gDemon
simplifies the implementation of GPU MMU virtualization by 37% code
reduction.

In detail, we have made the following changes in kernel.
* kernel parameters
    - shadow_execlist_context = 0
    - shadow_cmd_buffer = 0
* cancel write-protection for PPGT
* shift shadowed GGTT entry to a dediated high memory range
    - 128G offset
* add switch_gpu_iopt hypercall
    - locates at function vgt_do_render_context_switch
* add lookup_gpu_addr hypercall
    - in order to get vtd mapping for gfn
* dubug helper
    - dump lrca ring buffer and batch buffer in execlist
    - manually walk through gpu and vtd page table